Proposed CSS change to allow full-width <li/> for speaker with long bio.
Hey - not 100% convinced about this one... options are merge it, ignore it, or don't allow any speakers to have a bio > 100 characters?

I'm not keen on option (3) 'cos I think Louise Paling's bio is genuinely funny BECAUSE it's so long... and I don't much like the way it breaks the layout, so figured a full-width LI was probably the most elegant solution. But open to suggestions if you got a better idea?